requirements.txt
Django==3.1.5
pandas==1.2.0
djangorestframework==3.12.2
